Calling cleanUp() mutiple times allows numberOfHandlers to go negative.
This affects shutting down the background thread which checks for zero
value (indicating last background thread is exiting) to cause it to
cleanup global matters.

A number of other defensive changes were then made in order to force the
twitter4j streaming APIs to be under some kind of API control to the
consuming system.

The changes in this patch cause NPE to occur in the background thread,
however at least the background thread is under control now and will
exit.

The project really needs the background thread thing to be improved
and refactored.  Such as lifecycle events firing in a consistent way,
having control on stop/start, the replacement of the handler instance
ensuring the old handler is cleaned up.